dags/requirements.txt
pyOpenSSL==21.0.0
PyJWT==2.3.0
pandas<1.4,>=0.17.1
requests==2.27.1
requests-oauthlib==1.3.1
fireblocks-sdk
snowflake-connector-python==2.7.4
psycopg2-binary==2.9.3
xlrd==2.0.1
boto3==1.21.3
web3==5.28.0
ciso8601==2.2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
blockfrost-python==0.4.3
pysftp
asn1crypto==1.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
idna==3.3
oscrypto==1.2.1
pycparser==2.21
pycryptodomex==3.14.0
pytz==2021.3
six==1.16.0
urllib3==1.26.8
pyarrow==6.0.1